+++ composer/src/generate/tones.rs
@@ -11,9 +11,9 @@ impl RNG {
     }
 }
 
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 
-pub fn pulse50(frequency: f64, position: f64) -> f64 {
+pub fn pulse50(frequency: f32, position: f32) -> f32 {
     if frequency * position % 1.0 < 0.5 {
         1.0
     } else {
@@ -21,7 +21,7 @@ pub fn pulse50(frequency: f64, position: f64) -> f64 {
     }
 }
 
-pub fn pulse25(frequency: f64, position: f64) -> f64 {
+pub fn pulse25(frequency: f32, position: f32) -> f32 {
     if frequency * position % 1.0 < 0.5 {
         1.0
     } else {
@@ -29,7 +29,7 @@ pub fn pulse25(frequency: f64, position: f64) -> f64 {
     }
 }
 
-pub fn pulse125(frequency: f64, position: f64) -> f64 {
+pub fn pulse125(frequency: f32, position: f32) -> f32 {
     if frequency * position % 1.0 < 0.125 {
         1.0
     } else {
@@ -37,20 +37,20 @@ pub fn pulse125(frequency: f64, position: f64) -> f64 {
     }
 }
 
-pub fn triangle(f: f64, t: f64) -> f64 {
+pub fn triangle(f: f32, t: f32) -> f32 {
     2.0 / PI * (2.0 * PI * f * t).sin().asin()
 }
 
-pub fn saw(f: f64, t: f64) -> f64 {
+pub fn saw(f: f32, t: f32) -> f32 {
     2.0 * (f * (t + 0.5) % 1.0) - 1.0
 }
 
-pub fn sine(f: f64, t: f64) -> f64 {
+pub fn sine(f: f32, t: f32) -> f32 {
     (2.0 * PI * f * t).sin()
 }
 
 static mut NOISE_RNG: RNG = RNG(12345);
 
-pub fn noise(_: f64, _: f64) -> f64 {
-    unsafe { NOISE_RNG.next() as f64 / std::u64::MAX as f64 * 2.0 - 1.0 }
+pub fn noise(_: f32, _: f32) -> f32 {
+    unsafe { NOISE_RNG.next() as f32 / std::u64::MAX as f32 * 2.0 - 1.0 }
 }
